// bankauto/marketplace-ui/src/types/Service.ts
export enum ServiceStep {
  WORK_TYPE = 'workType',
  CHECK_VEHICLE_NUMBER = 'checkVehicleNumber',
  BRAND = 'brand',
  MODEL = 'model',
  YEAR = 'year',
  CAR_INFO = 'carInfo',
  SEARCH_MAP = 'searchMap',
  CONFIRMATION = 'confirmation',
}

export enum VehicleType {
  PASSENGER = 1,
  COMMERCIAL = 5,
}